requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-crontab==0.7.1
django-globals==0.3.2
django-import-export==1.2.0
django-stubs==0.9.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
djangorestframework-stubs==0.4.0
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.1
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
tablib==0.13.0
typed-ast==1.3.1
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
